chore!: move eslint-plugin-prettier to peerDependencies
BREAKING CHANGE:
This commit also moves eslint & prettier to peerDependencies.

After this change, users will have to manually install these
dependencies into their project root.

Previously we listed the plugin as a dependency of this config, and that
only works because package managers will hoist the plugin to project
root. This trick is not reliable with current eslint implementation.
This also resulted in a bug when the package manager failed to correctly
hoist the plugin package, see https://github.com/vuejs/vue-cli/issues/4310

More details can be seen at:
- https://github.com/eslint/eslint/issues/3458
- https://github.com/eslint/rfcs/pull/5